requirements.txt
streamlit==1.28.0
Pillow==10.1.0
ruff==0.1.3
requests==2.31.0
streamlit-lottie==0.0.5
watchdog==3.0.0
streamlit-option-menu==0.3.6